+++ tests/testthat/test-python-lists.R
@@ -35,3 +35,16 @@ test_that("Python unnamed tuples become R unnamed lists", {
 
   expect_equal(length(tuple(l)), length(l))
 })
+
+
+test_that("length method for Python lists works", {
+  skip_if_no_python()
+  py <- import_builtins(convert = FALSE)
+  l <- py$list()
+  l$append(1)
+  l$append(2)
+  l$append(3)
+  expect_equal(length(l), 3)
+})
+
+