+++ NEWS.md
@@ -21,6 +21,9 @@ Install the development version with: `install_github("rstudio/reticulate")`
 
 - Added `py_id()` function to get globally unique ids for Python objects
 
+- Added `py_len()` function and S3 `length()` method for Python lists (already
+  had `length()` methods for dicts, tuples, and NumPy arrays).
+
 
 ## reticulate 1.2 (CRAN)